change jimmer

+++ src/main/java/com/mellivorines/generator/service/GeneratorService.java
@@ -4,7 +4,7 @@ import java.util.zip.ZipOutputStream;
 
 public interface GeneratorService {
 
-    void downloadCode(Long tableId, ZipOutputStream zip);
+    void downloadCode(Integer tableId, ZipOutputStream zip);
 
     void generatorCode(Long tableId);
 }
